Remove unused solution platforms
The x64 platform didn't actually build most of Bonsai, it's an artifact of fb65c18a1960059fb0d913c4233b41fb9208fbfc when Bonsai64 was added (before becoming the default and leading to Bonsai32.)
The "Mixed Platforms" platform is something Visual Studio used to automatically add to solutions when multiple projects had conflicting platforms.

Technically Mixed Platforms was the default before, but Any CPU was equivalent and makes more sense in the context of Bonsai.