Move Directory.Build.* files into their own folder
This is being done in preparation for having auxiliary files for organization purposes.

The csproj-specific files pattern is intended to avoid issues when projects of multiple types share the same repo.
This isn't strictly necessary in the case of Bonsai, but I did it out of habit.

This also adds a dummy Directory.Build.rsp (another lesser-used ambient MSBuild configuration mechanism) to avoid searching for one outside of the repo.