Ensure member selector prefers derived properties
The issue with accessing inherited hidden properties also applies to the
member selector operator. In this case the accessor is for properties or
fields rather than just properties, but the same exact issues arise.

Again, we rewrite the default expression accessor to ensure properties
and fields are searched incrementally through the base type hierarchy.