Avoid multiple shutdown paths out of StartWorkflow
Here we improve the error routing when starting a workflow. Any workflow
errors are routed via the observable sequence either during the build
process or during subscription.

In both cases, or when successfully completing the sequence, the editor
running state must be cleared, so we track a single shared variable
which is disposed when exiting via any of these paths.