Ensure editor settings are saved on the right file
Saving on the form close handler is too late, as this can be triggered
when a new workflow has been opened in a separate folder, causing
the editor settings to be incorrectly saved to the new location.

A better moment is on close workflow, which may be triggered with
or without form reloading.