Ensure editor workflow is not set until validation
Opening a new workflow requires clearing a number of validation stages
following the initial load from file. Failing any of these stages
requires aborting the load and reverting back to the current workflow.
Here we store the candidate in a separate local until the entire
validation and workflow preparation is complete.