Avoid indirect loading of cyclic include workflows
The compiler performs validation and testing of include cycles as part
of the build. However, there is an edge case where it is still possible
to indirectly load a cyclic workflow by accessing its argument range.

These changes are to prevent such access purely for purposes of drawing
the include node range marks. This was causing the editor to allow
manual recursion into nested self-referential workflows. It will be safe
to remove this workaround when accessing the argument range of an
include workflow no longer triggers loading.