Allow showing visualizer window from context menu
The new VisualizerWindow operator does not allow changing the visualizer
type at runtime. However, it should be possible to relaunch the window
from the context menu in case it is closed.

To allow this we extend the visualizer context menu item with a click
action to launch the visualizer for the VisualizerWindow operator.