Model background tasks as observable sequences
Implementing each event handler as an observable sequence we can
do a joint subscribe on the cancellation token and avoid explicitly
awaiting a task. Since all sequences are observed on the form scheduler
this ensures exceptions are marshaled and recovered at the top-level.

The toolbox and workflow initialization are kept as tasks but awaited
in the main synchronization context to ensure any unhandled exceptions
can be recovered.