Avoid emitting redundant or invalid license URL
For backward compatibility purposes the builder will emit a license URL
field even when license metadata has been specified. However, if we
include it when we read back the metadata it will lead to an invalid
export since license metadata and license URL cannot both be used at
the same time.