Prune unnecessary recursion in predecessor search
When searching for predecessor inputs, we are either in the correct
workflow level, or in a parent level if the visualizer has been
propagated. In the latter case, we need to do recursive search.

However, we only need to recurse if the nested node represents either
the visualizer itself, or is displaying the visualizer as an overlay due
to visualizer mappings impinging on the node. Both tests can be readily
deduced from the existing API so we can effectively restrict recursion
only to the path containing the visualizer of interest.