Fallback to predecessor of propagated visualizer
In legacy visualizers we allowed overriding the predecessor when the
visualizer was propagated out from a nested node. This allowed for
example providing a background image in which to overlay a rectangle,
among other possibilities.

To avoid changing existing semantics, we allow finding the predecessor
from the propagated node only if no external input is acceptable.