Specify main workflow in gallery package metadata
Which specific workflow to launch was previously inferred by convention,
requiring an exact match between the file name and the package ID. This
introduced several difficulties during both export and import and was
ultimately limiting the evolution of gallery packages.

Here we introduce a new `bonsai.json` metadata file which will contain
the explicit declaration of which workflow to launch. We add both import
and export hooks in the gallery package infrastructure to leverage this
new specification.