Allow opening local environments on double-click
To improve the experience of double-clicking a workflow file in the
context of a local environment, we introduce a new launcher application
which automatically scans and bootstraps the most relevant environment.

The 32-bit launcher is removed as this would make the experience
confusing and is not considered relevant anymore. Note the default
launcher remains compatible with 32-bit operating systems, we simply
remove the shortcut to force launch in 32-bit when running in a 64-bit
system.